:bug: fix test and ci
