+++ .github/workflows/ci.yml
@@ -25,4 +25,4 @@ jobs:
       - name: Dependencies
         run: go get -v -t -d ./...
       - name: Go test
-        run: go test -v
+        run: go test -v ./...
